♻  refatorando de modal p/ tela - transformando modal de configuração de estatísticas em uma tela
