+++ src/navigation/PurchaseStack.js
@@ -6,6 +6,7 @@ import ProductOfListDeatils from '../screens/ProductOfListDetails/ProductOfListD
 import HistoryScreen from '../screens/HistoryScreen/HistoryScreen';
 import PurchaseDetailScreen from '../screens/PurchaseDetailScreen/PurchaseDetailScreen';
 import StatisticsScreen from '../screens/StatisticsScreen/StatisticsScreen';
+import StatisticsSettingsScreen from '../screens/StatisticsSettingsScreen/StatisticsSettingsScreen';
 
 import { useTranslation } from 'react-i18next';
 
@@ -54,6 +55,14 @@ export default function PurchaseStack() {
         name="Statistics"
         options={{ title: t('statistics'), headerStyle: stackHeaderStyle }}
         component={StatisticsScreen}
+      ></Stack.Screen>
+      <Stack.Screen
+        name="StatisticsSettings"
+        options={{
+          title: t('statisticsSettings'),
+          headerStyle: stackHeaderStyle,
+        }}
+        component={StatisticsSettingsScreen}
       ></Stack.Screen>
     </Stack.Navigator>
   );
