+++ src/screens/StatisticsScreen/StatisticsScreen.js
@@ -1,281 +1,57 @@
 /* eslint-disable no-case-declarations */
-import React, { useState, useContext, useEffect } from 'react';
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { SafeAreaView } from 'react-native';
-import {
-  Button,
-  Text,
-  HStack,
-  Box,
-  VStack,
-  Icon,
-  useToast,
-  View,
-} from 'native-base';
+import { Button, Text, HStack, Box, VStack, Icon, View } from 'native-base';
 import moment from 'moment';
 
-import { AuthContext } from '../../context/AuthProvider';
-import { ListContext } from '../../context/ListProvider';
-import StatisticsService from '../../services/StatisticsService';
 import { screenBasicStyle as style } from '../../styles/style';
+import { useFocusEffect } from '@react-navigation/native';
 import { Ionicons } from '@expo/vector-icons';
 import BarChartWrapper from '../../components/BarChartWrapper';
 import LineChartWrapper from '../../components/LineChartWrapper';
 
 import { useTranslation } from 'react-i18next';
-import StatisticsModal from '../../components/StatisticsModal';
-import CategoryService from '../../services/CategoryService';
-import ProductService from '../../services/ProductService';
 
-import {
-  UnityTimes,
-  DateParameters,
-  StatisticsType,
-  getUrl,
-} from '../../utils/StatisticsUtils';
+import { UnityTimes, StatisticsType } from '../../utils/StatisticsUtils';
 
-export default function StatisticsScreen() {
-  const [dataFromServer, setdataFromServer] = useState(null);
-  const [selectedUnityTime, setSelectedUnityTime] = useState('DAILY');
-  const [selectedStatisticsType, setSelectedStatisticsType] = useState(
-    StatisticsType.TIME
-  );
-  const [dateConfig, setDateConfig] = useState({
-    startDate: null,
-    endDate: null,
-  });
-  const [isSelectorOpen, setIsSelectorOpen] = useState(false);
-  const [currentParameter, setCurrentParameter] = useState(null);
-  const [isConfigOpen, setIsConfigOpen] = useState(false);
-  const [selectedList, setSelectedList] = useState(null);
-  const [categories, setCategories] = useState([]);
-  const [selectedCategory, setSelectedCategory] = useState(null);
-  const [products, setProducts] = useState([]);
-  const [selectedProduct, setSelectedProduct] = useState('');
-  const [loading, setLoading] = useState(false);
-  const [loadingCategories, setLoadingCategories] = useState(false);
-  const [loadingProducts, setLoadingProducts] = useState(false);
-
-  const { user } = useContext(AuthContext);
-  const { lists } = useContext(ListContext);
+export default function StatisticsScreen(props) {
   const { t } = useTranslation();
 
-  useEffect(() => {
-    if (!isConfigOpen && (!dateConfig.startDate || !dateConfig.endDate)) {
-      setDateConfig({
-        startDate: null,
-        endDate: null,
-      });
-    }
-  }, [isConfigOpen]);
+  const [dataFromServer, setdataFromServer] = useState(null);
+  const [statisticsSettings, setStatisticsSettings] = useState({
+    startDate: null,
+    endDate: null,
+    selectedUnityTime: UnityTimes.DAILY,
+    statisticType: StatisticsType.TIME,
+  });
 
-  // Quando o tipo de estatística selecionada for CATEGORY ou PRODUCT modifica programaticamente
-  // o selectedUnityTime para DEFAULT, pois esses tipos de análises não usam DAILY, WEEKLY nem MONTHLY e
-  // portanto podem utilizar o DatePicker no modo padrão
-  useEffect(() => {
+  const [statisticsName, setStatisticsName] = useState(
+    t(statisticsSettings.statisticType)
+  );
+
+  useFocusEffect(() => {
+    // Verifica se alguma tela enviou props para essa (a tela de config de estatísticas manda para cá)
     if (
-      selectedStatisticsType === StatisticsType.CATEGORY ||
-      selectedStatisticsType === StatisticsType.PRODUCT
+      props.route?.params?.settings &&
+      props.route?.params?.dataFromServer &&
+      props.route?.params?.statisticsName
     ) {
-      setSelectedUnityTime(UnityTimes.DEFAULT);
+      // Caso a tela anterior tenha passado settings
+      setStatisticsSettings(props.route.params.settings);
+      setdataFromServer(props.route.params.dataFromServer);
+      setStatisticsName(props.route.params.statisticsName);
     }
-
-    if (selectedStatisticsType === StatisticsType.CATEGORY) {
-      getCategories();
-    }
-  }, [selectedStatisticsType]);
-
-  const getStatisticsData = async () => {
-    let request;
-    let periodFilterObject;
-    try {
-      if (selectedStatisticsType !== StatisticsType.PURCHASE_LOCAL) {
-        periodFilterObject = {
-          startDate: moment(dateConfig.startDate).toISOString(),
-          endDate: moment(dateConfig.endDate).toISOString(),
-        };
-      }
-
-      console.log(periodFilterObject);
-
-      setLoading(true);
-      switch (selectedStatisticsType) {
-        case StatisticsType.PURCHASE_LOCAL:
-          request = StatisticsService.getPurchaseLocalData(user);
-          break;
-
-        case StatisticsType.TIME:
-          periodFilterObject = {
-            ...periodFilterObject,
-            unityTime: selectedUnityTime,
-          };
-
-          request = StatisticsService.getExpensesPer(
-            getUrl(selectedStatisticsType),
-            periodFilterObject,
-            user
-          );
-          break;
-
-        case StatisticsType.LIST:
-          periodFilterObject = {
-            ...periodFilterObject,
-            unityTime: selectedUnityTime,
-            listId: selectedList,
-          };
-          request = StatisticsService.getExpensesPer(
-            getUrl(selectedStatisticsType),
-            periodFilterObject,
-            user
-          );
-          break;
-
-        case StatisticsType.CATEGORY:
-          periodFilterObject = {
-            minDate: moment(dateConfig.startDate).toISOString(),
-            maxDate: moment(dateConfig.endDate).toISOString(),
-          };
-
-          periodFilterObject.category = categories.find(
-            (cat) => cat.id === selectedCategory
-          )?.name;
-
-          request = StatisticsService.getExpensesPer(
-            getUrl(selectedStatisticsType),
-            periodFilterObject,
-            user
-          );
-          break;
-
-        case StatisticsType.PRODUCT:
-          periodFilterObject.name = selectedProduct;
-          request = StatisticsService.getExpensesPer(
-            getUrl(selectedStatisticsType),
-            periodFilterObject,
-            user
-          );
-          break;
-
-        default:
-          break;
-      }
-
-      const { data } = await request;
-      console.log(data);
-      setdataFromServer(data);
-    } catch (error) {
-      console.log({ error });
-      useToast().show({
-        title: t('errorServerDefault'),
-        status: 'error',
-      });
-    } finally {
-      setLoading(false);
-      setIsConfigOpen(false);
-    }
-  };
-
-  const getCategories = async () => {
-    try {
-      setLoadingCategories(true);
-      const { data } = await CategoryService.getCategories(user);
-      setCategories(data);
-    } catch (error) {
-      console.log({ error });
-      useToast().show({
-        title: t('errorServerDefault'),
-        status: 'error',
-      });
-    } finally {
-      setLoadingCategories(false);
-    }
-  };
-
-  const handleMonthChange = (date) => {
-    if (currentParameter === DateParameters.START) {
-      setDateConfig({
-        ...dateConfig,
-        startDate: date.startOf('date'), // garante que pegará o dia definido desde às 00h00min00sec
-      });
-    } else {
-      setDateConfig({
-        ...dateConfig,
-        endDate: moment(date).endOf('month').endOf('date'), // garante que pegará o dia definido até as 23h59min59sec
-      });
-    }
-    setIsSelectorOpen(false);
-  };
-
-  const handleDailyChange = (date, dailyCurrentParameter) => {
-    if (dailyCurrentParameter === DateParameters.START) {
-      setDateConfig({
-        ...dateConfig,
-        startDate: moment(date).startOf('date'), // garante que pegará o dia definido desde às 00h00min00sec
-      });
-    } else {
-      setDateConfig({
-        ...dateConfig,
-        endDate: moment(date).endOf('date'), // garante que pegará o dia definido até as 23h59min59sec
-      });
-      setIsSelectorOpen(false);
-    }
-  };
-
-  const handleDefaultChange = (date) => {
-    if (currentParameter === DateParameters.START) {
-      setDateConfig({
-        ...dateConfig,
-        startDate: moment(date).startOf('date'), // garante que pegará o dia definido desde às 00h00min00sec
-      });
-    } else {
-      setDateConfig({
-        ...dateConfig,
-        endDate: moment(date).endOf('date'), // garante que pegará o dia definido até as 23h59min59sec
-      });
-    }
-    setIsSelectorOpen(false);
-  };
-
-  const handleWeeklyChange = (date) => {
-    if (currentParameter === DateParameters.START) {
-      setDateConfig({
-        ...dateConfig,
-        startDate: moment(date).startOf('date'), // garante que pegará o dia definido desde às 00h00min00sec
-      });
-    } else {
-      setDateConfig({
-        ...dateConfig,
-        endDate: moment(date).endOf('isoWeek').endOf('date'), // garante que pegará o dia definido até as 23h59min59sec
-      });
-    }
-    setIsSelectorOpen(false);
-  };
-
-  const handleDateChange = (date, currentParam = null) => {
-    switch (selectedUnityTime) {
-      case UnityTimes.DAILY:
-        handleDailyChange(date, currentParam);
-        break;
-
-      case UnityTimes.WEEKLY:
-        handleWeeklyChange(date);
-        break;
-
-      case UnityTimes.MONTHLY:
-        handleMonthChange(date);
-        break;
-
-      default:
-        handleDefaultChange(date);
-        break;
-    }
-  };
+  });
 
   const renderDateInterval = () => {
     let intervalText;
-    if (dateConfig.startDate && dateConfig.endDate) {
-      intervalText = `${moment(dateConfig.startDate).format('DD/MM/yyyy')} ${t(
-        'until'
-      )} ${moment(dateConfig.endDate).format('DD/MM/yyyy')}`;
+    if (statisticsSettings.startDate && statisticsSettings.endDate) {
+      intervalText = `${moment(statisticsSettings.startDate).format(
+        'DD/MM/yyyy'
+      )} ${t('until')} ${moment(statisticsSettings.endDate).format(
+        'DD/MM/yyyy'
+      )}`;
     } else {
       intervalText = t('noIntervalSelected');
     }
@@ -283,35 +59,15 @@ export default function StatisticsScreen() {
     return <Text>{intervalText}</Text>;
   };
 
-  const searchProducts = async (value) => {
-    if (value.length > 2) {
-      try {
-        setLoadingProducts(true);
-        const { data } = await ProductService.getProductByName(value, user);
-        console.log(data);
-        setProducts(data);
-      } catch (error) {
-        useToast().show({
-          title: t('errorServerDefault'),
-          status: 'warning',
-        });
-      } finally {
-        setLoadingProducts(false);
-      }
-    } else {
-      setProducts([]);
-    }
-  };
-
   const renderChart = () => {
-    switch (selectedStatisticsType) {
+    switch (statisticsSettings.statisticType) {
       case StatisticsType.PURCHASE_LOCAL:
         return;
       case StatisticsType.TIME:
       case StatisticsType.LIST:
         return (
           <BarChartWrapper
-            selectedUnityTime={selectedUnityTime}
+            selectedUnityTime={statisticsSettings.selectedUnityTime}
             monetaryNotation={t('currency')}
             preFormattedData={dataFromServer}
             translate={t}
@@ -339,40 +95,6 @@ export default function StatisticsScreen() {
     }
   };
 
-  const getStatisticsName = () => {
-    let chosenStatisticsString = t(selectedStatisticsType);
-
-    switch (selectedStatisticsType) {
-      case StatisticsType.CATEGORY:
-        if (selectedCategory) {
-          const categoryName = categories.find(
-            (cat) => cat.id === selectedCategory
-          )?.name;
-          chosenStatisticsString += `- ${categoryName}`;
-        }
-        break;
-
-      case StatisticsType.LIST:
-        if (selectedList) {
-          const listName = lists.find(
-            (list) => list.id === selectedList
-          )?.nameList;
-
-          chosenStatisticsString += `- ${listName}`;
-        }
-        break;
-
-      case StatisticsType.PRODUCT:
-        if (selectedProduct) chosenStatisticsString += `- ${selectedProduct}`;
-        break;
-
-      default:
-        break;
-    }
-
-    return chosenStatisticsString;
-  };
-
   return (
     <SafeAreaView style={style.container}>
       <View width="90%" mx="auto">
@@ -384,13 +106,17 @@ export default function StatisticsScreen() {
                   {t('chosenStatistics')}
                 </Text>
                 <Text fontSize="lg" fontWeight="bold" textAlign="center">
-                  {getStatisticsName()}
+                  {statisticsName}
                 </Text>
               </VStack>
 
               <Button
                 variant="ghost"
-                onPress={() => setIsConfigOpen(true)}
+                onPress={() =>
+                  props.navigation.navigate('StatisticsSettings', {
+                    settings: statisticsSettings,
+                  })
+                }
                 startIcon={
                   <Icon
                     size="sm"
@@ -404,14 +130,23 @@ export default function StatisticsScreen() {
             {/*  Caso o tipo de estatísticas seja uma relacionada com datas (qualquer uma que não seja a PURCHASE_LOCAL)
             e não houver um intervalo de datas selecionado, mostra uma mensagem indicando isso e um botão
           */}
-            {selectedStatisticsType !== StatisticsType.PURCHASE_LOCAL &&
-            !dateConfig.startDate &&
-            !dateConfig.endDate ? (
+            {statisticsSettings.statisticType !==
+              StatisticsType.PURCHASE_LOCAL &&
+            !statisticsSettings.startDate &&
+            !statisticsSettings.endDate ? (
               <Box mt={3}>
                 <Text fontSize="sm" textAlign="center" mr={5}>
                   {renderDateInterval()}
                 </Text>
-                <Button mt={2} size="sm" onPress={() => setIsConfigOpen(true)}>
+                <Button
+                  mt={2}
+                  size="sm"
+                  onPress={() =>
+                    props.navigation.navigate('StatisticsSettings', {
+                      settings: statisticsSettings,
+                    })
+                  }
+                >
                   {t('selectOne')}
                 </Button>
               </Box>
@@ -420,9 +155,10 @@ export default function StatisticsScreen() {
             {/*  Caso o tipo de estatísticas seja uma relacionada com datas (qualquer uma que não seja a PURCHASE_LOCAL)
             e tenha o intervalo de datas
           */}
-            {selectedStatisticsType !== StatisticsType.PURCHASE_LOCAL &&
-            dateConfig.startDate &&
-            dateConfig.endDate ? (
+            {statisticsSettings.statisticType !==
+              StatisticsType.PURCHASE_LOCAL &&
+            statisticsSettings.startDate &&
+            statisticsSettings.endDate ? (
               <Box>
                 <Text fontSize="sm" textAlign="center" mr={5}>
                   {renderDateInterval()}
@@ -433,43 +169,15 @@ export default function StatisticsScreen() {
         </VStack>
 
         {dataFromServer &&
-          dateConfig.startDate &&
-          dateConfig.endDate &&
+          statisticsSettings.startDate &&
+          statisticsSettings.endDate &&
           renderChart()}
-
-        <StatisticsModal
-          isConfigOpen={isConfigOpen}
-          setIsConfigOpen={setIsConfigOpen}
-          setSelectedStatisticsType={setSelectedStatisticsType}
-          selectedStatisticsType={selectedStatisticsType}
-          selectedUnityTime={selectedUnityTime}
-          setSelectedUnityTime={setSelectedUnityTime}
-          setDateConfig={setDateConfig}
-          translate={t}
-          setCurrentParameter={setCurrentParameter}
-          currentParameter={currentParameter}
-          renderDateInterval={renderDateInterval}
-          getStatisticsData={getStatisticsData}
-          handleDateChange={handleDateChange}
-          dateConfig={dateConfig}
-          isSelectorOpen={isSelectorOpen}
-          setIsSelectorOpen={setIsSelectorOpen}
-          selectedList={selectedList}
-          setSelectedList={setSelectedList}
-          lists={lists}
-          loading={loading}
-          categories={categories}
-          setSelectedCategory={setSelectedCategory}
-          selectedCategory={selectedCategory}
-          loadingCategories={loadingCategories}
-          searchProducts={searchProducts}
-          products={products}
-          setSelectedProduct={setSelectedProduct}
-          selectedProduct={selectedProduct}
-          setProducts={setProducts}
-          loadingProducts={loadingProducts}
-        />
       </View>
     </SafeAreaView>
   );
 }
+
+StatisticsScreen.propTypes = {
+  route: PropTypes.object,
+  navigation: PropTypes.object,
+};
